[#1853] - Mueve el ACL de LiteraryWork al repository (traduccion inline, sin mappers en _utils)
Cambio de direccion arquitectonica: la traduccion raw->dominio deja de vivir como
funciones de mapeo sueltas en src/api/_utils/literary-work.functions.ts y pasa a
ser responsabilidad del propio repository. El SanityLiteraryWorkRepository conoce
el shape de Sanity y entrega LiteraryWork listo; el service recibe dominio en el
formato que necesita, sin una capa de mappers aparte.

- La traduccion (mapWork, mapSectionProjection, mapMetadata, mapSection,
  mapEpigraph, toMaterializationInput) queda como metodos privados del adaptador
  Sanity. Reusa los sub-mappers transversales (mapAuthor/mapTags/... de _utils,
  compartidos con otros dominios) como building blocks.
- El doble InMemoryLiteraryWorkRepository pasa a almacenar dominio (LiteraryWork[])
  en vez de crudo: no replica el ACL, solo honra el contrato del puerto. Comparte
  projectSingleSection (proyeccion de dominio) con el adaptador real.
- Elimina literary-work.functions.ts y su spec; los tests de traduccion se folddean
  al spec del repository, ejercidos por la API publica (fetchBySlug/fetchSectionBySlug).

Diverge a proposito del patron fetch*()-crudo + mappers-en-_utils del resto del
backend, que no es la direccion objetivo de la plataforma.
