+++ src/api/modules/literary-work/literary-work.repository.mock.ts
@@ -1,20 +1,18 @@
 import type { LiteraryWork } from '@models/literary-work.model';
-import { mapLiteraryWork, type SanityLiteraryWork } from '../../_utils/literary-work.functions';
 import { projectSingleSection, type LiteraryWorkRepository } from './literary-work.repository';
 
-// Gemelo del SanityLiteraryWorkRepository: sustituye el almacenamiento (Sanity) por una lista en
-// memoria, pero honra el mismo contrato — guarda el shape de Sanity y mapea al dominio al leer.
-// No materializa (no hay CMS que persistir): el mapper deriva los reading time faltantes en el acto.
+// Doble de almacenamiento (Fake): sustituye Sanity por una lista en memoria de agregados de dominio ya
+// materializados. No traduce ni persiste — la anti-corruption es del adaptador Sanity; este solo honra el
+// contrato del puerto sirviendo dominio canned y proyectando secciones con el mismo helper compartido.
 export class InMemoryLiteraryWorkRepository implements LiteraryWorkRepository {
-	private readonly works: ReadonlyArray<SanityLiteraryWork>;
+	private readonly works: ReadonlyArray<LiteraryWork>;
 
-	constructor(works: ReadonlyArray<SanityLiteraryWork> = []) {
+	constructor(works: ReadonlyArray<LiteraryWork> = []) {
 		this.works = works;
 	}
 
 	public async fetchBySlug(slug: string): Promise<LiteraryWork | null> {
-		const raw = this.works.find((work) => work.slug === slug);
-		return raw ? mapLiteraryWork(raw) : null;
+		return this.works.find((work) => work.slug === slug) ?? null;
 	}
 
 	public async fetchSectionBySlug(slug: string, section: number): Promise<LiteraryWork | null> {
